// src/main/users/dto/createUser.dto.ts
export class CreateUser{
    userName?: string;
    profilePicture?: string;
    walletAddress?: string;
    publicKey?: string;
    aptosWalletAddress?: string;
    noOfConnectedAccounts?: number;
    isDiscordConnected?: boolean;
    isMetamaskConnected?: boolean;
    isPhantomConnected?: boolean;
    isAptosConnected?: boolean;
    isGoogleConnected?: boolean;
    discordId?: string;
    emailId?: string;
    profileName?: string;
    lastSignedIn?: Date;
    isWalletConnectConnected?: boolean;
    nearWalletKey?: boolean;
    isNearConnected?: boolean;
}